Makefile: draft epub target
There are still some issues with the epub this produces.
Most significantly, Pandoc for some reason sticks an empty section for
the Preface at the end of the file corresponding to the Title page.
Also, at least in the previews I look at in Calibre and with the
Firefox Epub reader plugin, the parallel layout is broken, and adding
a second stylesheet to turn off the German seems to have no effect.