TranslatorsNotes: remove extra parallel markup
I have so far been unable to find a *nice* way (i.e., not marking up
every paragraph with dummy .parallel divs and empty German blocks) of
displaying the whole section on the right hand side. This is due to a
combination of Pandoc's behavior and my lack of CSS knowledge. So I'm
leaving it out of the built version for now; I'll hopefully solve this
problem in 2.0