Ch II, parallel.css: an initial approach to equation numbering
use a <span> around equations with class "numbered", an id based on
page number, and data-label attribute containing the actual
label (since Hankel sometimes uses more than just numbers)

in CSS, display this <span> as a block with relative positioning,
and use a pseudoelement to display the label off to the side

This is still pretty manual (it requires me to enter all the labels by
hand, rather than using e.g. css counters), but seems to be pretty
clean and flexible enough to represent all the actual equation numbers
I've seen so far in the text.